+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - NEOM U
+### Fixed
+- Fix some styling issues
+- Hide HomePanel name info button
 
 ## [0.5.8] = 2025-02-15
 ### Added